--- src/main/java/duke/DukeException.java
@@ -1,14 +1,0 @@
-package duke;
-
-public class DukeException extends Exception {
-
-    protected String msg;
-    public DukeException(String msg) {
-        this.msg = msg;
-    }
-
-    @Override
-    public String toString() {
-        return this.msg;
-    }
-}
