+++ src/main/java/duke/commands/AddDeadlineCommand.java
@@ -1,10 +1,10 @@
 package duke.commands;
 
 import duke.Deadline;
-import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
 import duke.Ui;
+import duke.exceptions.DukeMissingArgumentException;
 
 public class AddDeadlineCommand extends Command {
 
@@ -13,14 +13,14 @@ public class AddDeadlineCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeMissingArgumentException {
         try {
             String[] str = description.substring(9).split(" /by ");
             Deadline deadline = new Deadline(str[0], str[1], false);
             tasks.add(deadline);
             ui.printAddTask(deadline, tasks.size());
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("OOPS!!! The description and/or the time of a deadline cannot be empty.");
+            throw new DukeMissingArgumentException("OOPS!!! The description and/or the time of a deadline cannot be empty.");
         }
     }
 }
