+++ src/main/java/duke/commands/AddEventCommand.java
@@ -1,10 +1,10 @@
 package duke.commands;
 
-import duke.DukeException;
 import duke.Event;
 import duke.Storage;
 import duke.TaskList;
 import duke.Ui;
+import duke.exceptions.DukeMissingArgumentException;
 
 public class AddEventCommand extends Command {
 
@@ -13,14 +13,14 @@ public class AddEventCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeMissingArgumentException {
         try {
             String[] str = description.substring(6).split(" /at ");
             Event event = new Event(str[0], str[1], false);
             tasks.add(event);
             ui.printAddTask(event, tasks.size());
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("OOPS!!! The description and/or the time of an event cannot be empty.");
+            throw new DukeMissingArgumentException("OOPS!!! The description and/or the time of an event cannot be empty.");
         }
     }
 }
