+++ src/main/java/duke/commands/AddToDoCommand.java
@@ -1,10 +1,10 @@
 package duke.commands;
 
-import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
 import duke.ToDo;
 import duke.Ui;
+import duke.exceptions.DukeMissingArgumentException;
 
 public class AddToDoCommand extends Command {
 
@@ -13,13 +13,13 @@ public class AddToDoCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeMissingArgumentException {
         try {
             ToDo todo = new ToDo(description.substring(5), false);
             tasks.add(todo);
             ui.printAddTask(todo, tasks.size());
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("OOPS!!! The description of a todo cannot be empty.");
+            throw new DukeMissingArgumentException("OOPS!!! The description of a todo cannot be empty.");
         }
     }
 }
