+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,10 +1,10 @@
 package duke.commands;
 
-import duke.DukeException;
 import duke.Storage;
 import duke.Task;
 import duke.TaskList;
 import duke.Ui;
+import duke.exceptions.DukeIndexOutOfBoundsException;
 
 public class DeleteCommand extends Command {
 
@@ -13,14 +13,14 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeIndexOutOfBoundsException {
         try {
             int index = Integer.parseInt(description.substring(7)) - 1;
             Task task = tasks.get(index);
             tasks.delete(index);
             ui.printDeleteTask(task, tasks.size());
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("OOPS!!! You cannot delete a non-existent task.");
+            throw new DukeIndexOutOfBoundsException("OOPS!!! You cannot delete a non-existent task.");
         }
     }
 }
