+++ src/main/java/duke/commands/MarkCommand.java
@@ -1,9 +1,9 @@
 package duke.commands;
 
-import duke.DukeException;
 import duke.Storage;
 import duke.TaskList;
 import duke.Ui;
+import duke.exceptions.DukeIndexOutOfBoundsException;
 
 public class MarkCommand extends Command {
 
@@ -12,13 +12,13 @@ public class MarkCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeIndexOutOfBoundsException {
         try {
             int index = Integer.parseInt(description.substring(5)) - 1;
             tasks.markTaskAsDone(index);
             ui.printTaskMarkedDone(tasks.get(index), tasks.size());
         } catch (IndexOutOfBoundsException e) {
-            throw new DukeException("OOPS!!! You cannot mark a non-existent task as done.");
+            throw new DukeIndexOutOfBoundsException("OOPS!!! You cannot mark a non-existent task as done.");
         }
     }
 }
