+++ src/main/java/duke/exceptions/DukeIndexOutOfBoundsException.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class DukeIndexOutOfBoundsException extends DukeException {
+
+    public DukeIndexOutOfBoundsException(String msg) {
+        super(msg);
+    }
+
+}
