+++ src/main/java/duke/exceptions/DukeMissingArgumentException.java
@@ -1,0 +1,9 @@
+package duke.exceptions;
+
+public class DukeMissingArgumentException extends DukeException {
+
+    public DukeMissingArgumentException(String msg) {
+        super(msg);
+    }
+
+}
